Add schema for UVTT format 0.2
Where possible, fields are considered optional even if we expect them to be present in all UVTT files. E.g., the various
array fields are optional since consumers can easily handle the missing case the same as an empty array. If a default
value is not applicable, alternative semantics are provided for when the field is missing.

The schema is specific to 0.2 UVTT files and thus the `"format"` field is pinned to `0.2`.